Update SohImGuiImpl.cpp
Probably fixed linux build error

+++ libultraship/libultraship/SohImGuiImpl.cpp
@@ -540,7 +540,7 @@ namespace SohImGui {
     void EnhancementCombo(const std::string& name, const char* cvarName, std::vector<std::string> items, int defaultValue) {
       
         if (ImGui::BeginCombo(name.c_str(), items[static_cast<int>(CVar_GetS32(cvarName, defaultValue))].c_str())) {
-            for (int settingIndex = 0; settingIndex < items.size(); settingIndex++) {
+            for (int settingIndex = 0; settingIndex < (int) items.size(); settingIndex++) {
                 if (ImGui::Selectable(items[settingIndex].c_str())) {
                     CVar_SetS32(cvarName, settingIndex);
                     needs_save = true;
